Fix lossy function result cast warnings
Add explicit static_cast<> for narrowing conversions to prevent
compiler warnings about lossy casts:

- TLuaInterpreter.cpp: Cast lua_tonumber (double) to qint64/int for
  direction numbers and color components
- TLuaInterpreterMudletObjects.cpp: Cast lua_tonumber to int for
  expiry counts and bool for flags
- TLuaInterpreterMapper.cpp: Cast lua_tonumber to int for area IDs
  and color components
- TMap.cpp: Use QRectF::toRect() for QPixmap::copy() which expects
  int parameters
- TCommandLine.cpp: Cast QSizeF::height() to int for line count
- T2DMap.cpp: Cast fabs() result to int for QPixmap constructor
- modern_glwidget.cpp: Cast QVector3D components (float) to int for
  map center coordinates

All casts are safe as the values are validated before use or the
truncation is intentional (e.g., map coordinates, color values 0-255).

+++ src/TCommandLine.cpp
@@ -661,7 +661,7 @@ void TCommandLine::adjustHeight()
         qWarning() << "TCommandLine::adjustHeight() ERROR: mpConsole->layerCommandLine is NULL!";
         return;
     }
-    int lines = document()->size().height();
+    int lines = static_cast<int>(document()->size().height());
     // Workaround for SubCommandLines textCursor not visible in some situations
     // SubCommandLines cannot autoresize
     if (mType == SubCommandLine) {
